+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Use [Lerna](https://lernajs.io/) to create the @converse/headless package
 - Use ES2015 modules instead of UMD.
+- #1257: Prefer 'probably' over 'maybe' when evaluating audio play support.
 
 ## 4.0.3 (2018-10-22)